# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/currency/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Currency"
                ],
                "summary": "Get list of supported currencies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quotation.GetCurrencyListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotation/last-requested": {
            "get": {
                "description": "Retrieves last requested quotation by base and quote currencies. Returns `404 Quotation not found` if quotation wasn't requested at least once, use `[post] /api/v1/update-request` in this case",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quotation"
                ],
                "summary": "Get last requested quotation by currencies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base Currency",
                        "name": "base",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Quote Currency",
                        "name": "quote",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quotation.GetQuotationResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Quotation not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotation/update-request": {
            "post": {
                "description": "Creates a quotation update request. Returns request Id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quotation"
                ],
                "summary": "Request quotation update",
                "parameters": [
                    {
                        "description": "Quotation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quotation.RequestQuotationUpdateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quotation.RequestQuotationUpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/quotation/update-request/{id}": {
            "get": {
                "description": "Retrieves a quotation by request Id. If request is not proceeded yet, returns status `NotReady`. If request is completed, returns status `Ready` and fields `rate` and `updatedAt`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quotation"
                ],
                "summary": "Get quotation by request Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quotation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quotation.GetQuotationByRequestIdResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No request with such id",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "quotation.GetCurrencyListResponse": {
            "type": "object",
            "properties": {
                "currencies": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            }
        },
        "quotation.GetQuotationByRequestIdResponse": {
            "description": "fields `price` and `timestamp` are only presented when status is `Ready`",
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "rate": {
                    "type": "string",
                    "format": "decimal",
                    "example": "123.45"
                },
                "status": {
                    "$ref": "#/definitions/quotation.RequestStatus"
                },
                "updatedAt": {
                    "description": "Unix timestamp in milliseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1694613600
                }
            }
        },
        "quotation.GetQuotationResponse": {
            "type": "object",
            "properties": {
                "rate": {
                    "type": "string",
                    "format": "decimal",
                    "example": "123.45"
                },
                "updatedAt": {
                    "description": "Unix timestamp in milliseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1694613600
                }
            }
        },
        "quotation.RequestQuotationUpdateBody": {
            "type": "object",
            "required": [
                "baseCurrency",
                "idempotencyKey",
                "quoteCurrency"
            ],
            "properties": {
                "baseCurrency": {
                    "$ref": "#/definitions/types.Currency"
                },
                "idempotencyKey": {
                    "type": "string",
                    "format": "uuid"
                },
                "quoteCurrency": {
                    "$ref": "#/definitions/types.Currency"
                }
            }
        },
        "quotation.RequestQuotationUpdateResponse": {
            "type": "object",
            "required": [
                "requestId"
            ],
            "properties": {
                "requestId": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "quotation.RequestStatus": {
            "type": "string",
            "enum": [
                "Ready",
                "NotReady"
            ],
            "x-enum-varnames": [
                "Ready",
                "NotReady"
            ]
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "types.Currency": {
            "type": "string",
            "enum": [
                "USD",
                "EUR",
                "MXN"
            ],
            "x-enum-varnames": [
                "USD",
                "EUR",
                "MXN"
            ]
        }
    }
}