# internal/domain/types/currency.go
package types

import "encoding/json"

type Currency string

const (
	USD Currency = "USD"
	EUR Currency = "EUR"
	MXN Currency = "MXN"
)

func AllCurrencies() []Currency {
	return []Currency{USD, EUR, MXN}
}

func (c Currency) IsValid() bool {
	switch c {
	case USD, EUR, MXN:
		return true

	default:
		return false
	}
}

func (c Currency) MarshalJSON() ([]byte, error) {
	return json.Marshal(string(c))
}